fix Review model

+++ api_yamdb/reviews/models.py
@@ -62,7 +62,7 @@ class Review(models.Model):
         on_delete=models.CASCADE,
         related_name='reviews',
     )
-    score = models.IntegerField(
+    score = models.PositiveSmallIntegerField(
         default=1,
         validators=[
             MaxValueValidator(10),
